+++ src/main/java/org/ajay/data/task/Todo.java
@@ -1,6 +1,6 @@
-package org.ajay.task;
+package org.ajay.data.task;
 
-import org.ajay.exceptions.EmptyArgumentException;
+import org.ajay.data.exceptions.EmptyArgumentException;
 
 public class Todo extends Task {
     public final static String COMMAND_STRING = "todo"; // Command string for the Todo class
